% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignAndCombine.R
\name{alignAndCombine}
\alias{alignAndCombine}
\title{alignAndCombine Function}
\usage{
alignAndCombine(
  reference,
  gff1,
  gff2,
  time.it = T,
  quiet = F,
  filenum1 = "1",
  filenum2 = "2",
  seqA = 1,
  seqB = 2
)
}
\arguments{
\item{reference}{The file reference positons}

\item{gff1}{The gff file which is the reference gff}

\item{gff2}{the gff file that needs rearranging}

\item{filenum1}{number or name for the first gff file}

\item{filenum2}{number or name for the second gff file}
}
\description{
alignAndCombine Function
}
\examples{
alignAndCombine()
}
\keyword{combines}
\keyword{different}
\keyword{files}
\keyword{from}
\keyword{genomes}
\keyword{gff}
\keyword{multiple}
