% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacterialCodonsFormat.R
\name{bacterialCodonsFormat}
\alias{bacterialCodonsFormat}
\title{bacterialCodonsFormat Function}
\usage{
bacterialCodonsFormat(bac = bacterial)
}
\arguments{
\item{bac}{The bacterial file formatted by bacterialColumnNames}
}
\description{
bacterialCodonsFormat Function
}
\examples{
bacterialCodonsFormat()
}
\keyword{bacterial}
\keyword{data}
\keyword{reformat}
\keyword{the}
