% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacterialColumnNames.R
\name{bacterialColumnNames}
\alias{bacterialColumnNames}
\title{bacterialColumnNames Function}
\usage{
bacterialColumnNames(bac = bacterial)
}
\arguments{
\item{bac}{The input bacterial file}
}
\description{
bacterialColumnNames Function
}
\examples{
bacterialColumnNames()
}
\keyword{bacterial}
\keyword{data}
\keyword{reformat}
\keyword{the}
