% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacterialDinucleotideFormat.R
\name{bacterialDinucleotideFormat}
\alias{bacterialDinucleotideFormat}
\title{bacterialDinucleotideFormat Function}
\usage{
bacterialDinucleotideFormat(bac = bacterial)
}
\arguments{
\item{bac}{The bacterial file formatted by bacterialColumnNames}
}
\description{
bacterialDinucleotideFormat Function
}
\examples{
bacterialDinucleotideFormat()
}
\keyword{bacterial}
\keyword{data}
\keyword{reformat}
\keyword{the}
