% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReferenceLookup.R
\name{buildReferenceLookup}
\alias{buildReferenceLookup}
\title{buildReferenceLookup Function}
\usage{
buildReferenceLookup(
  reference,
  seqA = 1,
  seqB = 2,
  remove.missing = T,
  collapse.alignment = T,
  quiet = F
)
}
\arguments{
\item{reference}{The file reference positons}

\item{seqA}{which columns to use for the first sequence}

\item{seqB}{which columns to use for the second sequence}

\item{remove.missing}{remove rows that are not conserved}

\item{collapse.alignment}{combines segments which can be combines once other alignments are removed}
}
\description{
buildReferenceLookup Function
}
\examples{
buildReferenceLookup()
}
\keyword{+ve}
\keyword{.backbone}
\keyword{Takes}
\keyword{a}
\keyword{all}
\keyword{and}
\keyword{file}
\keyword{full}
\keyword{info}
\keyword{input}
\keyword{produces}
\keyword{sequences,}
\keyword{sorted}
\keyword{strand}
\keyword{two}
\keyword{values}
\keyword{with}
