% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmscanToGff.R
\name{cmscanToGff}
\alias{cmscanToGff}
\title{cmscanToGff Function}
\usage{
cmscanToGff(rfamRes)
}
\arguments{
\item{rfamRes}{The cmscan output file}
}
\description{
cmscanToGff Function
}
\examples{
cmscanToGff()
}
\keyword{Convert}
\keyword{This}
\keyword{a}
\keyword{cmscan}
\keyword{file}
\keyword{file.}
\keyword{formatted}
\keyword{gff}
\keyword{header}
\keyword{include}
\keyword{info.}
\keyword{not}
\keyword{output}
\keyword{the}
\keyword{to}
\keyword{top}
\keyword{will}
