% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonSummary.R
\name{codonSummary}
\alias{codonSummary}
\title{codonSummary Function}
\usage{
codonSummary(ph = phageCodons, bac = bacCodons, codon_order = F)
}
\arguments{
\item{ph}{The phage dataset formated by phageCodonsFormat}

\item{bac}{The bacterial dataset formated by bacterialCodonsFormat}

\item{codon_order}{Determines whether to reorder the codons by amino acid frequency or not}
}
\description{
codonSummary Function
}
\examples{
codonSummary()
}
\keyword{data}
\keyword{phage}
\keyword{reformat}
\keyword{the}
