% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseAlignment.R
\name{collapseAlignment}
\alias{collapseAlignment}
\title{collapseAlignment Function}
\usage{
collapseAlignment(ref)
}
\arguments{
\item{ref}{The file reference positons}
}
\description{
collapseAlignment Function
}
\examples{
collapseAlignment()
}
\keyword{Takes}
\keyword{a}
\keyword{alignment}
\keyword{an}
\keyword{and}
\keyword{any}
\keyword{contain}
\keyword{file}
\keyword{merges}
\keyword{ongoing}
\keyword{reference}
\keyword{rows}
\keyword{sequences}
\keyword{that}
\keyword{two}
\keyword{with}
