% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAASummary.R
\name{createAASummary}
\alias{createAASummary}
\title{createAASummary Function}
\usage{
createAASummary(aa = aminoAcidFreq, cl = codonLookup)
}
\arguments{
\item{aa}{The amino acid frequency in the bacterial genomes}

\item{cl}{Codons mapping to amino acids}
}
\description{
createAASummary Function
}
\examples{
createAASummary()
}
\keyword{data}
\keyword{phage}
\keyword{reformat}
\keyword{the}
