% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinucleotideSummary.R
\name{dinucleotideSummary}
\alias{dinucleotideSummary}
\title{dinucleotideSummary Function}
\usage{
dinucleotideSummary(ph = phageDinucleotides, bac = bacDinucleotides)
}
\arguments{
\item{ph}{The phage dataset formated by phageDinucleotideFormat}

\item{bac}{The bacterial dataset formated by bacterialDinucleotideFormat}
}
\description{
dinucleotideSummary Function
}
\examples{
dinucleotideSummary()
}
\keyword{data}
\keyword{phage}
\keyword{reformat}
\keyword{the}
