% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreeLengths.R
\name{getTreeLengths}
\alias{getTreeLengths}
\title{getTreeLengths Function}
\usage{
getTreeLengths(treeLengths, treeDat, name)
}
\arguments{
\item{treeLengths}{list containing tree lengths}

\item{treeDat}{data frame with node numbers and lengths}

\item{name}{name of branch for extracting from list}
}
\description{
getTreeLengths Function
}
\examples{
getTreeLengths()
}
\keyword{gets}
\keyword{lengths}
\keyword{tip.label}
