% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_all.R
\name{interset_all}
\alias{interset_all}
\title{interset_all Function}
\usage{
interset_all(set_val)
}
\arguments{
\item{set_val}{The file reference positons}
}
\description{
interset_all Function
}
\examples{
mergeSRA()
}
\keyword{all}
\keyword{and}
\keyword{finds}
\keyword{interrsect}
\keyword{mutiple}
\keyword{of}
\keyword{set}
\keyword{takes}
\keyword{the}
\keyword{them}
\keyword{values}
