% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mauvePlotColourChange.R
\name{mauvePlotColourChange}
\alias{mauvePlotColourChange}
\title{mauvePlotColourChange Function}
\usage{
mauvePlotColourChange(bbone)
}
\arguments{
\item{bbone}{the genoPlotR mauve file}
}
\description{
mauvePlotColourChange Function
}
\examples{
mauvePlotColourChange()
}
\keyword{bbone}
\keyword{change}
\keyword{colors}
\keyword{file}
\keyword{on}
\keyword{the}
