% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSRA.R
\name{mergeSRA}
\alias{mergeSRA}
\title{mergeSRA Function}
\usage{
mergeSRA(
  ncRNAgff,
  gff1,
  gff2,
  time.it = T,
  quiet = F,
  filenum1 = "1",
  filenum2 = "2",
  initial_data = T,
  align = T
)
}
\arguments{
\item{gff1}{The gff file which is the reference gff}

\item{gff2}{the gff file that needs rearranging}

\item{filenum1}{number or name for the first gff file}

\item{filenum2}{number or name for the second gff file}

\item{ref}{The file reference positons}
}
\description{
mergeSRA Function
}
\examples{
mergeSRA()
}
\keyword{combines}
\keyword{different}
\keyword{files}
\keyword{from}
\keyword{genomes}
\keyword{gff}
\keyword{multiple}
