% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phageCodonsFormat.R
\name{phageCodonsFormat}
\alias{phageCodonsFormat}
\title{phageCodonsFormat Function}
\usage{
phageCodonsFormat(ph = phage)
}
\arguments{
\item{ph}{The phage file formatted by phageColumnNames}
}
\description{
phageCodonsFormat Function
}
\examples{
phageCodonsFormat()
}
\keyword{data}
\keyword{phage}
\keyword{reformat}
\keyword{the}
