% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phageColumnNames.R
\name{phageColumnNames}
\alias{phageColumnNames}
\title{phageColumnNames Function}
\usage{
phageColumnNames(ph = phage)
}
\arguments{
\item{bac}{The input phage file}
}
\description{
phageColumnNames Function
}
\examples{
phageColumnNames()
}
\keyword{data}
\keyword{phage}
\keyword{reformat}
\keyword{the}
