% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phageDinucleotideFormat.R
\name{phageDinucleotideFormat}
\alias{phageDinucleotideFormat}
\title{phageDinucleotideFormat Function}
\usage{
phageDinucleotideFormat(ph = phage)
}
\arguments{
\item{ph}{The phage file formatted by phageColumnNames}
}
\description{
phageDinucleotideFormat Function
}
\examples{
phageDinucleotideFormat()
}
\keyword{data}
\keyword{phage}
\keyword{reformat}
\keyword{the}
