% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCodons.R
\name{plotCodons}
\alias{plotCodons}
\title{plotCodons Function}
\usage{
plotCodons(cc = codons, output = "plot")
}
\arguments{
\item{cc}{The codon dataset}

\item{output}{plot or write the graph}
}
\description{
plotCodons Function
}
\examples{
plotCodons()
}
\keyword{data}
\keyword{phage}
\keyword{reformat}
\keyword{the}
