% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDinuclteotides.R
\name{plotDinucleotides}
\alias{plotDinucleotides}
\title{plotDinucleotides Function}
\usage{
plotDinucleotides(cc = dinucleotides, output = "plot")
}
\arguments{
\item{cc}{The dinucleotide dataset}

\item{output}{plot or write the graph}
}
\description{
plotDinucleotides Function
}
\examples{
plotDinucleotides()
}
\keyword{data}
\keyword{phage}
\keyword{reformat}
\keyword{the}
