% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGC.R
\name{plotGC}
\alias{plotGC}
\title{plotGC Function}
\usage{
plotGC(ph = phage, tr = tRNA, cd = CDS, output = "plot")
}
\arguments{
\item{ph}{phage dataset}

\item{tr}{bacterial tRNAs}

\item{cd}{bacterial CDS}

\item{output}{plot or write the graph}
}
\description{
plotGC Function
}
\examples{
plotGC()
}
\keyword{data}
\keyword{phage}
\keyword{reformat}
\keyword{the}
