% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCDSregions.R
\name{removeCDSregions}
\alias{removeCDSregions}
\title{removeCDSregions Function}
\usage{
removeCDSregions(
  plotDat,
  gffDat,
  buffer = 150,
  quiet = F,
  stranded = F,
  time.it = F
)
}
\arguments{
\item{plotDat}{The file with the plot values}

\item{gffDat}{The gff file that contains the CDS locations}
}
\description{
removeCDSregions Function
}
\examples{
removeCDSregions()
}
\keyword{Convert}
\keyword{This}
\keyword{a}
\keyword{cmscan}
\keyword{file}
\keyword{file.}
\keyword{formatted}
\keyword{gff}
\keyword{header}
\keyword{include}
\keyword{info.}
\keyword{not}
\keyword{output}
\keyword{the}
\keyword{to}
\keyword{top}
\keyword{will}
