% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderGFF.R
\name{reorderGFF}
\alias{reorderGFF}
\title{reorderGFF Function}
\usage{
reorderGFF(ref, gff, time.it = T, quiet = F, reference.genome = F)
}
\arguments{
\item{ref}{The file reference positons}

\item{gff}{the gff file that needs rearranging}
}
\description{
reorderGFF Function
}
\examples{
reorderGFF()
}
\keyword{a}
\keyword{file}
\keyword{for}
\keyword{gff}
\keyword{positions}
\keyword{relabels}
