% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderSraPlot.R
\name{reorderSraPlot}
\alias{reorderSraPlot}
\title{reorderSraPlot Function}
\usage{
reorderSraPlot(
  sra,
  alignment,
  reference,
  seqA = 1,
  seqB = 2,
  quiet = F,
  time.it = T
)
}
\arguments{
\item{sra}{the sra file to rearrange}

\item{alignment}{the alignment file}

\item{reference}{The reference sra file}
}
\description{
reorderSraPlot Function
}
\examples{
reorderSraPlot()
}
\keyword{Reorder}
\keyword{allow}
\keyword{be}
\keyword{compared}
\keyword{file}
\keyword{genomes}
\keyword{mutplie}
\keyword{plot}
\keyword{sra}
\keyword{the}
\keyword{to}
