% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaPeakCalling.R
\name{rnaPeakCalling}
\alias{rnaPeakCalling}
\title{rnaPeakCalling Function}
\usage{
rnaPeakCalling(
  dat,
  col.num = 1,
  quiet = F,
  time.it = T,
  plot_threshold = 10,
  small_peaks = T,
  binwidth = 25
)
}
\arguments{
\item{dat}{The plot file to smooth and sort}

\item{col.num}{1 for forward strand and 2 for reverse strand}

\item{binwidth}{the width to smooth over}

\item{moveby}{the jumps to make for each smoothing bin}
}
\description{
rnaPeakCalling Function
}
\examples{
rnaPeakCalling()
}
\keyword{and}
\keyword{expression}
\keyword{file,}
\keyword{find}
\keyword{is}
\keyword{occuring.}
\keyword{out}
\keyword{plot}
\keyword{ranges}
\keyword{smooth}
\keyword{the}
\keyword{where}
