% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreRNAReadDepth.R
\name{scoreRNAReadDepth}
\alias{scoreRNAReadDepth}
\title{scoreRNAReadDepth Function}
\usage{
scoreRNAReadDepth(file_path, gff_names, sra, allData)
}
\arguments{
\item{file_path}{location of all the files to work with}

\item{gff_names}{location of the gff files relative to the file path}

\item{sra}{the gff file to use}

\item{allData}{the file to append the data to}
}
\description{
scoreRNAReadDepth Function
}
\examples{
scoreRNAReadDepth()
}
\keyword{RNASeq}
\keyword{a}
\keyword{all}
\keyword{and}
\keyword{calls}
\keyword{depths}
\keyword{files}
\keyword{from}
\keyword{gets}
\keyword{in}
\keyword{list}
\keyword{of}
\keyword{read}
\keyword{take}
\keyword{the}
