% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sraSmoothAndSort.R
\name{sraSmoothandsort}
\alias{sraSmoothandsort}
\title{sraSmoothandsort Function}
\usage{
sraSmoothandsort(dat, col.num = 1, binwidth = 25, moveby = 5)
}
\arguments{
\item{dat}{The plot file to smooth and sort}

\item{col.num}{1 for forward strand and 2 for reverse strand}

\item{binwidth}{the width to smooth over}

\item{moveby}{the jumps to make for each smoothing bin}
}
\description{
sraSmoothandsort Function
}
\examples{
sraSmoothandsort()
}
\keyword{and}
\keyword{expression}
\keyword{file,}
\keyword{find}
\keyword{is}
\keyword{occuring.}
\keyword{out}
\keyword{plot}
\keyword{ranges}
\keyword{smooth}
\keyword{the}
\keyword{where}
